refactor(rust): rename AgentRegistry -> AgentTable, move source to services rlib
Architectural placement: agent lifecycle state is a service concern, not
a kernel primitive. Source lives in the services rlib alongside Python
nexus.services.agents, mirroring the rest of the services tier. The
PyO3 binding shell stays in the kernel cdylib because nexus_kernel is
the single Python-facing extension module (rust/raft/Cargo.toml documents
why two cdylibs cannot share a Python process namespace; AgentTable
follows the same pattern raft already established).

Moves:
- rust/kernel/src/agent_registry.rs (deleted) -> rust/services/src/agent_table.rs
- AgentRegistry -> AgentTable (Linux task_struct array analogue, distinct
  name from the Python AgentRegistry user-facing service)
- AgentDescriptor / AgentState / AgentKind / AgentObserver move with it;
  fields go pub (was pub(crate)) so the kernel cdylib can re-export them
- AgentStatusResolver split into its own kernel file
  rust/kernel/src/agent_status_resolver.rs - the resolver implements a
  kernel-internal PathResolver trait so it stays in the kernel crate,
  reading via Arc<services::agent_table::AgentTable>

Wiring:
- rust/services/Cargo.toml gains dashmap / parking_lot / serde / serde_json
- rust/kernel/Cargo.toml depends on the services rlib
- Kernel struct field renamed agent_registry -> agent_table
- Codegen template + generated_pyo3.rs use services::agent_table::* paths
- Stubs regenerated; Python-facing kernel.agent_* surface unchanged

Tests: services::agent_table::tests (10 tests) + agent_status_resolver test
all pass.
