+++ rust/kernel/Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["cdylib", "rlib"]  # cdylib for Python, rlib for Rust deps (e.g. r
 [dependencies]
 contracts = { workspace = true }
 library = { workspace = true }
+services = { workspace = true }
 # F2 C8 (Option A): raft rlib pulled in so kernel can impl Metastore over
 # raft's ZoneConsensus directly, no cross-cdylib Rust trait object
 # dispatch. See rust/kernel/src/raft_metastore.rs.
