+++ rust/services/src/lib.rs
@@ -1,2 +1,4 @@
 //! Rust service implementations.
 //! Aligned with Python `src/nexus/services/`.
+
+pub mod agent_table;
